app/requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
exceptiongroup==1.2.0
fastapi==0.108.0
ffmpeg==1.4
filelock==3.13.1
fonttools==4.44.3
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
Jinja2==3.1.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.41.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper==20231117
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
referencing==0.32.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.32.0.post1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
pytest==7.4.4